Make xcl_import_binary xclbin discovery work with SDx GUI projects.
This enhances the xclbin discovery to look in a list of typical xclbin
directories, for a list of typical xclbin filename patterns, to match
command-line build filename patterns, and a few SDx GUI filename
patterns. This allows both command line and gui flows to properly load
the xclbin files without needing to modify the host code.